+++ static/libs/code.js
@@ -1,1 +1,1 @@
-export{U as init}from"./code-BW5_X4Xh.js";
+export{U as init}from"./code-wUp8fBb_.js";