+++ static/libs/markdown.js
@@ -1,1 +1,1 @@
-export{w as init}from"./markdown-B-XyGvZG.js";
+export{w as init}from"./markdown-TMAtxxgD.js";